refactor better api requests folders
