+++ app/lib/backend/http/api/messages.dart
@@ -1,0 +1,52 @@
+import 'dart:convert';
+
+import 'package:flutter/material.dart';
+import 'package:friend_private/backend/http/shared.dart';
+import 'package:friend_private/backend/schema/message.dart';
+import 'package:friend_private/env/env.dart';
+
+Future<List<ServerMessage>> getMessagesServer() async {
+  // TODO: Add pagination
+  var response = await makeApiCall(url: '${Env.apiBaseUrl}v1/messages', headers: {}, method: 'GET', body: '');
+  if (response == null) return [];
+  debugPrint('getMessages: ${response.body}');
+  if (response.statusCode == 200) {
+    var messages =
+        (jsonDecode(response.body) as List<dynamic>).map((memory) => ServerMessage.fromJson(memory)).toList();
+    debugPrint('getMessages length: ${messages.length}');
+    return messages;
+  }
+  return [];
+}
+
+Future<ServerMessage> sendMessageServer(String text, {String? pluginId}) {
+  return makeApiCall(
+    url: '${Env.apiBaseUrl}v1/messages?plugin_id=$pluginId',
+    headers: {},
+    method: 'POST',
+    body: jsonEncode({'text': text}),
+  ).then((response) {
+    if (response == null) throw Exception('Failed to send message');
+    if (response.statusCode == 200) {
+      return ServerMessage.fromJson(jsonDecode(response.body));
+    } else {
+      throw Exception('Failed to send message');
+    }
+  });
+}
+
+Future<ServerMessage> getInitialPluginMessage(String? pluginId) {
+  return makeApiCall(
+    url: '${Env.apiBaseUrl}v1/initial-message?plugin_id=$pluginId',
+    headers: {},
+    method: 'POST',
+    body: '',
+  ).then((response) {
+    if (response == null) throw Exception('Failed to send message');
+    if (response.statusCode == 200) {
+      return ServerMessage.fromJson(jsonDecode(response.body));
+    } else {
+      throw Exception('Failed to send message');
+    }
+  });
+}
