+++ app/lib/backend/http/openai.dart
@@ -1,10 +1,55 @@
 import 'dart:convert';
 import 'dart:io';
+import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
-import 'package:friend_private/backend/api_requests/api/shared.dart';
+import 'package:friend_private/backend/http/shared.dart';
+import 'package:friend_private/backend/database/memory.dart';
 import 'package:friend_private/backend/preferences.dart';
+import 'package:friend_private/backend/schema/plugin.dart';
 import 'package:friend_private/env/env.dart';
+import 'package:tuple/tuple.dart';
+
+class SummaryResult {
+  final Structured structured;
+  final List<Tuple2<Plugin, String>> pluginsResponse;
+
+  SummaryResult(this.structured, this.pluginsResponse);
+}
+
+Future<String> triggerTestMemoryPrompt(String prompt, String transcript) async {
+  return await executeGptPrompt('''
+        Your are an AI with the following characteristics:
+        Task: $prompt
+        
+        Note: It is possible that the conversation you are given, has nothing to do with your task, \
+        in that case, output an empty string. (For example, you are given a business conversation, but your task is medical analysis)
+        
+        Conversation: ```${transcript.trim()}```,
+       
+        Output your response in plain text, without markdown.
+        Make sure to be concise and clear.
+        '''
+      .replaceAll('     ', '')
+      .replaceAll('    ', '')
+      .trim());
+}
+
+Future<String> getPhotoDescription(Uint8List data) async {
+  var messages = [
+    {
+      'role': 'user',
+      'content': [
+        {'type': "text", 'text': "What’s in this image?"},
+        {
+          'type': "image_url",
+          'image_url': {"url": "data:image/jpeg;base64,${base64Encode(data)}"},
+        },
+      ],
+    },
+  ];
+  return await gptApiCall(model: 'gpt-4o', messages: messages, maxTokens: 100);
+}
 
 Future<dynamic> gptApiCall({
   required String model,
