+++ app/lib/pages/capture/location_service.dart
@@ -49,7 +49,7 @@ class LocationService {
   Future<Geolocation?> getGeolocationDetails() async {
     if (await hasPermission()) {
       LocationData locationData = await location.getLocation();
-      // TODO: move http requests to other.dart
+      // TODO: move http requests to webhooks.dart
 
       try {
         var res = await http.get(
