+++ app/lib/pages/memory_detail/test_prompts.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:friend_private/backend/api_requests/api/prompt.dart';
-import 'package:friend_private/backend/database/memory.dart';
-import 'package:friend_private/backend/server/memory.dart';
+import 'package:friend_private/backend/http/openai.dart';
+import 'package:friend_private/backend/schema/memory.dart';
 
 class TestPromptsPage extends StatefulWidget {
   final ServerMemory memory;
