+++ app/lib/pages/speaker_id/tabs/completed.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:friend_private/backend/mixpanel.dart';
+import 'package:friend_private/utils/analytics/mixpanel.dart';
 import 'package:friend_private/backend/preferences.dart';
 
 class CompletionTab extends StatefulWidget {
