+++ app/lib/utils/memories/integrations.dart
@@ -1,6 +1,6 @@
-import 'package:friend_private/backend/api_requests/api/other.dart';
+import 'package:friend_private/backend/http/webhooks.dart';
 import 'package:friend_private/backend/database/transcript_segment.dart';
-import 'package:friend_private/backend/server/message.dart';
+import 'package:friend_private/backend/schema/message.dart';
 import 'package:friend_private/utils/other/notifications.dart';
 
 triggerTranscriptSegmentReceivedEvents(
